todo_project/requirements.txt
Django>=4.0
djangorestframework>=3.14
drf-yasg>=1.21